refactor: Apply Facade Pattern implementaion
For CustomCombatPanel class, create FacadeCustomCombatPanel class to apply Facade Pattern at FacadeCustomCombatPanel.java file
